+++ cmd/src/users.go
@@ -20,7 +20,7 @@ The commands are:
 	get        gets a user
 	create     creates a user account
 	delete     deletes a user account
-	clean      deletes inactive users
+	prune      deletes inactive users
 	tag        add/remove a tag on a user
 
 Use "src users [command] -h" for more information about a command.
@@ -89,4 +89,12 @@ type UserEmail struct {
 type UserUsageStatistics struct {
 	LastActiveTime                    string
 	LastActiveCodeHostIntegrationTime string
+}
+
+type SiteUser struct {
+	ID           string
+	Username     string
+	Email        string
+	SiteAdmin    bool
+	LastActiveAt string
 }